refactor: remove unneeded Property subclasses
